[#9] Add option for hashing individual files before uploading
Individual files were being uploaded with their original filenames. This is a nice piece of functionality to have, but I don't want it to be the default. There's now an option in the preferences that allows the user to determine if they want those files to have their names hashed when it gets uploaded.