fix: Stop uploading file multiple times
It was an invisible bug, but monitoring the Eidetica logs would reveal
that when a screenshot was taken on macOS, the app would attempt to
upload the screenshot multiple times. This is because the OS does
multiple rename operations on the file.

I've added a more stringent check on filenames to eliminate the first
couple of unnecessary upload attempts, as well as a screenshot queue to
track which files we've already uploaded.